% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_rules.R
\name{learn_rules}
\alias{learn_rules}
\title{Execution of rule induction algorithm.}
\usage{
learn_rules(formula, control, train_data, test_data = train_data)
}
\arguments{
\item{formula}{Formulae specifing label attribute and attributes included in rule induction (e.g. class ~ .)}

\item{control}{Named list with parameters for algorithm.}

\item{train_data}{Data frame or list of data frames with training data.}

\item{test_data}{Data frame or list of data frames with test data. If not specified is equals to train_data.}
}
\value{
List containing data frame with performance statistics for each data set and character vector with generated rules.
}
\description{
Execution of rule induction algorithm.
}
\examples{
control <- list(min_rule_covered = 11, induction_measure = 'C2', pruning_measure = 'C2', max_growing = 0)
train_data <- iris
formula <- Species ~ .

results <- learn_rules(formula, control, train_data)
}
