% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{methane_train}
\alias{methane_train}
\title{Training data for methane levle prediction}
\format{A data frame with 13368 rows and 8 variables}
\description{
Training data for methane levle prediction
}
\details{
Attribute information.
\itemize{
 \item{MM31}{numeric attribute}
 \item{MM116}{numeric attribute}
 \item{AS038}{numeric attribute}
 \item{PG072}{numeric attribute}
 \item{PD}{numeric attribute}
 \item{BA13}{numeric attribute}
 \item{DMM116}{numeric attribute}
 \item{MM116_pred}{label attribute}
 }
}
\keyword{datasets}
